+++ src/Autocomplete/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 -   Minimum required Symfony version is now 6.4
 -   Minimum required PHP version is now 8.2
+-   Remove `ParentEntityAutocompleteType` in favor of `BaseEntityAutocompleteType`
+-   Remove `ExtraLazyChoiceLoader` in favor of `Symfony\Component\Form\ChoiceList\Loader\LazyChoiceLoader` from Symfony Form >=7.2
+-   Add parameter `$security` to `AutocompleteResultsExecutor::__construct()`
 
 ## 2.30